+++ phoenix-scala/phoenix/app/phoenix/models/customer/CustomerRank.scala
@@ -1,15 +1,17 @@
 package phoenix.models.customer
 
+
 import shapeless._
 import slick.jdbc.PostgresProfile.api._
+import utils.Money._
 import utils.db._
 
-case class CustomerRank(id: Int = 0, revenue: Int = 0, rank: Option[Int] = Some(0))
+case class CustomerRank(id: Int = 0, revenue: Long = 0, rank: Option[Int] = Some(0))
     extends FoxModel[CustomerRank] {}
 
 class CustomersRanks(tag: Tag) extends FoxTable[CustomerRank](tag, "customers_search_view") {
   def id      = column[Int]("id", O.PrimaryKey)
-  def revenue = column[Int]("revenue")
+  def revenue = column[Long]("revenue")
   def rank    = column[Option[Int]]("rank")
 
   def * = (id, revenue, rank) <> ((CustomerRank.apply _).tupled, CustomerRank.unapply)
