+++ phoenix-scala/phoenix/app/phoenix/models/discount/offers/ItemsAmountOffer.scala
@@ -11,7 +11,7 @@ import phoenix.utils.aliases._
 import phoenix.utils.apis.Apis
 
 // Amount off all matched items in cart
-case class ItemsAmountOffer(discount: Int, search: Seq[ProductSearch])
+case class ItemsAmountOffer(discount: Long, search: Seq[ProductSearch])
     extends Offer
     with AmountOffer
     with NonEmptySearch
