+++ phoenix-scala/phoenix/app/phoenix/models/discount/qualifiers/OrderTotalAmountQualifier.scala
@@ -5,7 +5,7 @@ import phoenix.utils.aliases._
 import phoenix.utils.apis.Apis
 import utils.db._
 
-case class OrderTotalAmountQualifier(totalAmount: Int) extends Qualifier {
+case class OrderTotalAmountQualifier(totalAmount: Long) extends Qualifier {
 
   val qualifierType: QualifierType = OrderTotalAmount
