+++ phoenix-scala/phoenix/app/phoenix/models/shipping/ShippingMethod.scala
@@ -6,6 +6,7 @@ import phoenix.failures.ShippingMethodFailures.ShippingMethodIsNotActive
 import phoenix.models.cord.OrderShippingMethods
 import phoenix.models.rules.QueryStatement
 import shapeless._
+import utils.Money._
 import utils.db.ExPostgresDriver.api._
 import utils.db._
 
@@ -15,7 +16,7 @@ case class ShippingMethod(id: Int = 0,
                           storefrontDisplayName: String,
                           code: String,
                           shippingCarrierId: Option[Int] = None,
-                          price: Int,
+                          price: Long,
                           isActive: Boolean = true,
                           conditions: Option[QueryStatement] = None,
                           restrictions: Option[QueryStatement] = None)
@@ -46,7 +47,7 @@ class ShippingMethods(tag: Tag) extends FoxTable[ShippingMethod](tag, "shipping_
   def storefrontDisplayName = column[String]("storefront_display_name")
   def code                  = column[String]("code")
   def shippingCarrierId     = column[Option[Int]]("shipping_carrier_id")
-  def price                 = column[Int]("price")
+  def price                 = column[Long]("price")
   def isActive              = column[Boolean]("is_active")
   def conditions            = column[Option[QueryStatement]]("conditions")
   def restrictions          = column[Option[QueryStatement]]("restrictions")
