+++ phoenix-scala/phoenix/app/phoenix/responses/GiftCardAdjustmentsResponse.scala
@@ -2,11 +2,12 @@ package phoenix.responses
 
 import phoenix.models.payment.InStorePaymentStates
 import phoenix.models.payment.giftcard.GiftCardAdjustment
+import utils.Money._
 
 object GiftCardAdjustmentsResponse {
   case class Root(id: Int,
-                  amount: Int,
-                  availableBalance: Int,
+                  amount: Long,
+                  availableBalance: Long,
                   state: InStorePaymentStates.State,
                   cordRef: Option[String])
       extends ResponseItem
