+++ phoenix-scala/phoenix/app/phoenix/services/activity/CartTailored.scala
@@ -7,6 +7,7 @@ import phoenix.models.payment.PaymentMethod
 import phoenix.models.shipping.ShippingMethod
 import phoenix.responses.cord.CartResponse
 import phoenix.responses.{AddressResponse, CreditCardsResponse, GiftCardResponse, UserResponse}
+import utils.Money._
 
 object CartTailored {
 
@@ -70,19 +71,19 @@ object CartTailored {
 
   case class CartPaymentMethodAddedGiftCard(cart: CartResponse,
                                             giftCard: GiftCardResponse.Root,
-                                            amount: Int,
+                                            amount: Long,
                                             admin: Option[UserResponse.Root])
       extends ActivityBase[CartPaymentMethodAddedGiftCard]
 
   case class CartPaymentMethodUpdatedGiftCard(cart: CartResponse,
                                               giftCard: GiftCardResponse.Root,
-                                              oldAmount: Option[Int],
-                                              amount: Int,
+                                              oldAmount: Option[Long],
+                                              amount: Long,
                                               admin: Option[UserResponse.Root])
       extends ActivityBase[CartPaymentMethodUpdatedGiftCard]
 
   case class CartPaymentMethodAddedStoreCredit(cart: CartResponse,
-                                               amount: Int,
+                                               amount: Long,
                                                admin: Option[UserResponse.Root])
       extends ActivityBase[CartPaymentMethodAddedStoreCredit]
